+++ lib/src/_shared/ui/osu_ui.dart
@@ -1,5 +1,6 @@
 export 'osu_badges.dart';
 export 'player_card.dart';
+export 'affiliation_tile.dart';
 export 'beatmap_card.dart';
 export 'play_card.dart';
 export 'news_card.dart';